componentized page sections
